shuffle=None

+++ tets.py
@@ -287,7 +287,7 @@ class CusRandomSampler(Sampler):
 if __name__ == '__main__':
     # img1 = loader('face.jpg', 'extract').unsqueeze(0)
     # img2 = loader('face.jpg', 'extract').unsqueeze(0)
-    model = SiameseNetwork(False).to(device)
+    model = SiameseNetwork(False).to(device).half()
     #
     # # print(model.forward_bilinear(img1,img2).data.cpu().numpy())
     #
